Move MessageBuilder
