+++ adala/utils/llm_utils.py
@@ -11,32 +11,9 @@ from litellm.types.utils import Usage
 from tenacity import Retrying, AsyncRetrying
 from instructor.exceptions import InstructorRetryException, IncompleteOutputException
 from instructor.client import Instructor, AsyncInstructor
-from adala.utils.parse import partial_str_format, split_message_into_chunks, MessageChunkType
+from adala.utils.parse import MessagesBuilder
 
 logger = logging.getLogger(__name__)
-
-
-class MessagesBuilder(BaseModel):
-    user_prompt_template: str
-    system_prompt: Optional[str] = None
-    instruction_first: bool = True
-    extra_fields: Dict[str, Any] = Field(default_factory=dict)
-    split_into_chunks: bool = False
-    input_field_types: Optional[Dict[str, MessageChunkType]] = Field(default=None)
-
-    def get_messages(self, payload: Dict[str, Any]):
-        if self.split_into_chunks:
-            input_field_types = self.input_field_types or defaultdict(lambda: MessageChunkType.TEXT)
-            user_prompt = split_message_into_chunks(self.user_prompt_template, input_field_types, **payload, **self.extra_fields)
-        else:
-            user_prompt = partial_str_format(self.user_prompt_template, **payload, **self.extra_fields)
-        messages = [{"role": "user", "content": user_prompt}]
-        if self.system_prompt:
-            if self.instruction_first:
-                messages.insert(0, {"role": "system", "content": self.system_prompt})
-            else:
-                messages[0]["content"] += self.system_prompt
-        return messages
 
 
 def _get_usage_dict(usage: Usage, model: str) -> Dict:
